Better organise the checks file structure
