+++ lib/src/checks/checks.dart
@@ -415,149 +415,3 @@ class _GroupCheck extends Check {
     return result;
   }
 }
-
-/// A check that checks that the user that executes a command has a specific role.
-class RoleCheck extends Check {
-  /// The IDs of the roles this check allows.
-  Iterable<Snowflake> roleIds;
-
-  /// Create a new [RoleCheck] that succeeds if the user that created the context has [role].
-  ///
-  /// You might also be interested in:
-  /// - [RoleCheck.id], for creating this same check without an instance of [IRole];
-  /// - [RoleCheck.any], for checking that the user that created a context has one of a set or
-  ///   roles.
-  RoleCheck(IRole role, [String? name]) : this.id(role.id, name);
-
-  /// Create a new [RoleCheck] that succeeds if the user that created the context has a role with
-  /// the id [id].
-  RoleCheck.id(Snowflake id, [String? name])
-      : roleIds = [id],
-        super(
-          (context) => context.member?.roles.any((role) => role.id == id) ?? false,
-          name ?? 'Role Check on $id',
-        );
-
-  /// Create a new [RoleCheck] that succeeds if the user that created the context has any of [roles].
-  ///
-  /// You might also be interested in:
-  /// - [RoleCheck.anyId], for creating this same check without instances of [IRole].
-  RoleCheck.any(Iterable<IRole> roles, [String? name])
-      : this.anyId(roles.map((role) => role.id), name);
-
-  /// Create a new [RoleCheck] that succeeds if the user that created the context has any role for
-  /// which the role's id is in [roles].
-  RoleCheck.anyId(Iterable<Snowflake> roles, [String? name])
-      : roleIds = roles,
-        super(
-          (context) => context.member?.roles.any((role) => roles.contains(role.id)) ?? false,
-          name ?? 'Role Check on any of [${roles.join(', ')}]',
-        );
-}
-
-/// A check that checks that a command was executed by a specific user.
-class UserCheck extends Check {
-  /// The IDs of the users this check allows.
-  Iterable<Snowflake> userIds;
-
-  /// Create a new [UserCheck] that succeeds if the context was created by [user].
-  ///
-  /// You might also be interested in:
-  /// - [UserCheck.id], for creating this same check without an instance of [IUser],
-  /// - [UserCheck.any], for checking that a context was created by a user in a set or users.
-  UserCheck(IUser user, [String? name]) : this.id(user.id, name);
-
-  /// Create a new [UserCheck] that succeeds if the ID of the user that created the context is [id].
-  UserCheck.id(Snowflake id, [String? name])
-      : userIds = [id],
-        super((context) => context.user.id == id, name ?? 'User Check on $id');
-
-  /// Create a new [UserCheck] that succeeds if the context was created by any one of [users].
-  ///
-  /// You might also be interested in:
-  /// - [UserCheck.anyId], for creating this same check without instance of [IUser].
-  UserCheck.any(Iterable<IUser> users, [String? name])
-      : this.anyId(users.map((user) => user.id), name);
-
-  /// Create a new [UserCheck] that succeeds if the ID of the user that created the context is in
-  /// [ids].
-  UserCheck.anyId(Iterable<Snowflake> ids, [String? name])
-      : userIds = ids,
-        super(
-          (context) => ids.contains(context.user.id),
-          name ?? 'User Check on any of [${ids.join(', ')}]',
-        );
-}
-
-/// A check that checks that a command was executed in a particular guild, or in a channel that is
-/// not in a guild.
-///
-/// This check is special as commands with this check will only be registered as slash commands in
-/// the guilds specified by this guild check. For this functionality to work, however, this check
-/// must be a "top-level" check - that is, a check that is not nested within a modifier such as
-/// [Check.any], [Check.deny] or [Check.all].
-///
-/// The value of this check overrides [CommandsPlugin.guild].
-///
-/// This check integrates with the [Discord Slash Command Permissions](https://discord.com/developers/docs/interactions/application-commands#permissions)
-/// API, so users that cannot use a command because of this check will have that command appear
-/// unavailable out in their Discord client.
-///
-/// You might also be interested in:
-/// - [CommandsPlugin.guild], for globally setting a guild to register slash commands to.
-class GuildCheck extends Check {
-  /// The IDs of the guilds that this check allows.
-  ///
-  /// If [guildIds] is `[null]`, then any guild is allowed, but not channels outside of guilds.
-  Iterable<Snowflake?> guildIds;
-
-  /// Create a [GuildCheck] that succeeds if the context originated in [guild].
-  ///
-  /// You might also be interested in:
-  /// - [GuildCheck.id], for creating this same check without an instance of [IGuild];
-  /// - [GuildCheck.any], for checking if the context originated in any of a set of guilds.
-  GuildCheck(IGuild guild, [String? name]) : this.id(guild.id, name);
-
-  /// Create a [GuildCheck] that succeeds if the ID of the guild the context originated in is [id].
-  GuildCheck.id(Snowflake id, [String? name])
-      : guildIds = [id],
-        super((context) => context.guild?.id == id, name ?? 'Guild Check on $id', false);
-
-  /// Create a [GuildCheck] that succeeds if the context originated outside of a guild (generally,
-  /// in private messages).
-  ///
-  /// You might also be interested in:
-  /// - [GuildCheck.all], for checking that a context originated in a guild.
-  GuildCheck.none([String? name])
-      : guildIds = [],
-        super((context) => context.guild == null, name ?? 'Guild Check on <none>', true, 0);
-
-  /// Create a [GuildCheck] that succeeds if the context originated in a guild.
-  ///
-  /// You might also be interested in:
-  /// - [GuildCheck.none], for checking that a context originated outside a guild.
-  GuildCheck.all([String? name])
-      : guildIds = [null],
-        super(
-          (context) => context.guild != null,
-          name ?? 'Guild Check on <any>',
-          false,
-        );
-
-  /// Create a [GuildCheck] that succeeds if the context originated in any of [guilds].
-  ///
-  /// You might also be interested in:
-  /// - [GuildCheck.anyId], for creating the same check without instances of [IGuild].
-  GuildCheck.any(Iterable<IGuild> guilds, [String? name])
-      : this.anyId(guilds.map((guild) => guild.id), name);
-
-  /// Create a [GuildCheck] that succeeds if the id of the guild the context originated in is in
-  /// [ids].
-  GuildCheck.anyId(Iterable<Snowflake> ids, [String? name])
-      : guildIds = ids,
-        super(
-          (context) => ids.contains(context.guild?.id),
-          name ?? 'Guild Check on any of [${ids.join(', ')}]',
-          false,
-        );
-}
